+++ liana-gui/src/app/buysell/onramper.rs
@@ -1,0 +1,1 @@
+const API_KEY: &'static str = "pk_test_01K2HQVXK7F5C8RDZ36WV2W3F5";
