+++ liana-gui/src/app/mod.rs
@@ -19,28 +19,6 @@ use iced::{clipboard, time, widget::Column, Subscription, Task};
 use tokio::runtime::Handle;
 use tracing::{error, info, warn};
 
-// Ultralight webview imports
-#[cfg(feature = "webview")]
-use iced_webview::{Action as WebviewAction, PageType, Ultralight, WebView};
-
-// Separate message type for webview that implements Clone
-#[cfg(feature = "webview")]
-#[derive(Debug, Clone)]
-pub enum WebviewMessage {
-    Action(WebviewAction),
-    Created,
-    UrlChanged(String),
-}
-
-// Create optimized webview with performance settings
-#[cfg(feature = "webview")]
-fn init_webview<E: iced_webview::Engine + Default>() -> iced_webview::WebView<E, WebviewMessage>
-{
-    WebView::new()
-        .on_create_view(WebviewMessage::Created)
-        .on_url_change(WebviewMessage::UrlChanged)
-}
-
 pub use liana::miniscript::bitcoin;
 use liana_ui::{component::network_banner, widget::Element};
 pub use lianad::{commands::CoinStatus, config::Config as DaemonConfig};
@@ -79,7 +57,9 @@ struct Panels {
     create_spend: CreateSpendPanel,
     settings: SettingsState,
     #[cfg(feature = "dev-meld")]
-    meld_buy_and_sell: crate::app::view::meld_buysell::MeldBuySellPanel,
+    meld_buy_sell: crate::app::view::meld_buysell::BuySellPanel,
+    #[cfg(feature = "dev-onramp")]
+    onramper_buy_sell: crate::app::view::meld_buysell::BuySellPanel,
 }
 
 impl Panels {
@@ -133,7 +113,7 @@ impl Panels {
                 config.clone(),
             ),
             #[cfg(feature = "dev-meld")]
-            meld_buy_and_sell: crate::app::view::meld_buysell::MeldBuySellPanel::new(cache.network),
+            meld_buy_sell: crate::app::view::meld_buysell::BuySellPanel::new(cache.network),
         }
     }
 
@@ -151,7 +131,9 @@ impl Panels {
             Menu::RefreshCoins(_) => &self.create_spend,
             Menu::PsbtPreSelected(_) => &self.psbts,
             #[cfg(feature = "dev-meld")]
-            Menu::BuyAndSell => &self.meld_buy_and_sell,
+            Menu::BuySell => &self.meld_buy_sell,
+            #[cfg(not(feature = "dev-meld"))]
+            Menu::BuySell => panic!("BuySell feature disabled in this build"),
         }
     }
 
@@ -169,7 +151,9 @@ impl Panels {
             Menu::RefreshCoins(_) => &mut self.create_spend,
             Menu::PsbtPreSelected(_) => &mut self.psbts,
             #[cfg(feature = "dev-meld")]
-            Menu::BuyAndSell => &mut self.meld_buy_and_sell,
+            Menu::BuySell => &mut self.meld_buy_sell,
+            #[cfg(not(feature = "dev-meld"))]
+            Menu::BuySell => panic!("BuySell feature disabled in this build"),
         }
     }
 }
@@ -181,32 +165,6 @@ pub struct App {
     internal_bitcoind: Option<Bitcoind>,
 
     panels: Panels,
-
-    // Ultralight webview component for Meld widget integration with performance optimizations
-    #[cfg(feature = "webview")]
-    webview: Option<WebView<Ultralight, WebviewMessage>>,
-
-    // Flag to indicate when webview should be rendered instead of normal panels
-    webview_mode: bool,
-
-    // Flag to track webview loading state
-    webview_loading: bool,
-
-    // Flag to track if webview is ready to be rendered (view has been created)
-    webview_ready: bool,
-
-    // Flag to control whether webview widget is shown
-    show_webview: bool,
-
-    // Current webview URL for display
-    current_webview_url: Option<String>,
-
-    // Timestamp when webview loading started (for timeout detection)
-    webview_loading_start: Option<std::time::Instant>,
-
-    // Webview management fields (following iced_webview example pattern)
-    num_webviews: u32,
-    current_webview_index: Option<u32>,
 }
 
 impl App {
@@ -237,16 +195,6 @@ impl App {
                 daemon,
                 wallet,
                 internal_bitcoind,
-                #[cfg(feature = "webview")]
-                webview: None,
-                webview_mode: false,
-                webview_loading: false,
-                webview_ready: false,
-                show_webview: false,
-                current_webview_url: None,
-                webview_loading_start: None,
-                num_webviews: 0,
-                current_webview_index: None,
             },
             cmd,
         )
@@ -262,22 +210,8 @@ impl App {
                 return alias;
             }
         }
-        "Liana wallet"
-    }
 
-    /// Check if webview is currently loading
-    pub fn is_webview_loading(&self) -> bool {
-        self.webview_loading
-    }
-
-    /// Map webview messages to main app messages (static version for Task::map)
-    #[cfg(feature = "webview")]
-    fn map_webview_message_static(webview_msg: WebviewMessage) -> Message {
-        match webview_msg {
-            WebviewMessage::Action(action) => Message::View(view::Message::WebviewAction(action)),
-            WebviewMessage::Created => Message::View(view::Message::WebviewCreated),
-            WebviewMessage::UrlChanged(url) => Message::View(view::Message::WebviewUrlChanged(url)),
-        }
+        "Coincube Vault Wallet"
     }
 
     fn set_current_panel(&mut self, menu: Menu) -> Task<Message> {
@@ -356,7 +290,7 @@ impl App {
     }
 
     pub fn subscription(&self) -> Subscription<Message> {
-        let mut subscriptions = vec![
+        let subscriptions = [
             time::every(Duration::from_secs(
                 match sync_status(
                     self.daemon.backend(),
@@ -388,18 +322,6 @@ impl App {
             .map(|_| Message::Tick),
             self.panels.current().subscription(),
         ];
-
-        // Add webview update subscription for smooth rendering when webview is active
-        #[cfg(feature = "webview")]
-        if self.webview_mode && self.show_webview && self.webview_ready {
-            subscriptions.push(
-                time::every(Duration::from_secs(1)) // 1 second for faster loading detection and content updates
-                    .map(|_| {
-                        use iced_webview::Action as WebViewAction;
-                        Message::View(view::Message::WebviewAction(WebViewAction::Update))
-                    }),
-            );
-        }
 
         Subscription::batch(subscriptions)
     }
@@ -487,16 +409,7 @@ impl App {
                     Message::WalletUpdated(Ok(wallet)),
                 )
             }
-            Message::View(view::Message::Menu(menu)) => {
-                match menu {
-                    menu::Menu::BuyAndSell => {
-                        // Switch to buy/sell panel and show modal
-                        self.panels.current = menu;
-                        Task::none()
-                    }
-                    _ => self.set_current_panel(menu),
-                }
-            }
+            Message::View(view::Message::Menu(menu)) => self.set_current_panel(menu),
             Message::View(view::Message::OpenUrl(url)) => {
                 if let Err(e) = open::that_detached(&url) {
                     tracing::error!("Error opening '{}': {}", url, e);
@@ -505,203 +418,7 @@ impl App {
             }
             Message::View(view::Message::Clipboard(text)) => clipboard::write(text),
 
-            Message::View(view::Message::OpenWebview(url)) => {
-                tracing::info!("Opening widget URL in embedded webview");
-
-                // Load URL into Ultralight webview
-                #[cfg(feature = "webview")]
-                {
-                    tracing::info!("🌐 [LIANA] Loading Ultralight webview with URL: {}", url);
-                    self.webview_mode = true;
-                    self.webview_loading = true;
-                    self.webview_loading_start = Some(std::time::Instant::now());
-                    self.current_webview_url = Some(url.clone());
-
-                    // Create webview with URL string and immediately update to ensure content loads
-                    let webview = self.webview.get_or_insert_with(init_webview);
-                    let create_task = webview
-                        .update(WebviewAction::CreateView(PageType::Url(url)))
-                        .map(Self::map_webview_message_static);
-
-                    // Add immediate update to trigger content loading
-                    let immediate_update = Task::done(Message::View(view::Message::WebviewAction(
-                        WebviewAction::Update,
-                    )));
-
-                    Task::batch([create_task, immediate_update])
-                }
-                #[cfg(not(feature = "webview"))]
-                Task::none()
-            }
-            #[cfg(feature = "webview")]
-            Message::View(view::Message::WebviewAction(action)) => {
-                // Handle webview-only actions - does NOT trigger full app update
-                tracing::debug!("🌐 [LIANA] Processing webview-only action: {:?}", action);
-
-                use iced_webview::Action as WebViewAction;
-
-                // Determine if this action needs a follow-up update for rendering
-                let needs_update = matches!(
-                    action,
-                    WebViewAction::CreateView(_)
-                        | WebViewAction::Resize(_)
-                        | WebViewAction::GoToUrl(_)
-                );
-
-                let webview = self.webview.get_or_insert_with(init_webview);
-                let main_task = webview.update(action).map(Self::map_webview_message_static);
-
-                if needs_update {
-                    // Add a single update after actions that change content/size
-                    let update_task = Task::done(Message::View(view::Message::WebviewAction(
-                        WebViewAction::Update,
-                    )));
-                    Task::batch([main_task, update_task])
-                } else {
-                    main_task
-                }
-            }
-            #[cfg(feature = "webview")]
-            Message::View(view::Message::WebviewCreated) => {
-                tracing::info!("🌐 [LIANA] Webview created successfully");
-                self.webview_mode = true;
-                self.webview_loading = false;
-                self.webview_ready = true;
-
-                // Increment view count and switch to the first view (following iced_webview example pattern)
-                self.num_webviews += 1;
-
-                // Switch to the first view and immediately update to display content
-                let switch_task = Task::done(Message::View(view::Message::SwitchToWebview(0)));
-                let update_task = Task::done(Message::View(view::Message::WebviewAction(
-                    iced_webview::Action::Update,
-                )));
-                Task::batch([switch_task, update_task])
-            }
-            #[cfg(feature = "webview")]
-            Message::View(view::Message::SwitchToWebview(index)) => {
-                tracing::info!("🌐 [LIANA] Switching to webview index: {}", index);
-
-                // Update current view index in app state
-                self.current_webview_index = Some(index);
-
-                // Send ChangeView action to webview and immediately update to display content
-                let webview = self.webview.get_or_insert_with(init_webview);
-                let change_task = webview
-                    .update(WebviewAction::ChangeView(index))
-                    .map(Self::map_webview_message_static);
-                let update_task = Task::done(Message::View(view::Message::WebviewAction(
-                    WebviewAction::Update,
-                )));
-                Task::batch([change_task, update_task])
-            }
-            #[cfg(feature = "webview")]
-            Message::View(view::Message::WebviewUrlChanged(url)) => {
-                tracing::info!("🌐 [LIANA] Webview URL changed to: {}", url);
-                self.current_webview_url = Some(url);
-                Task::none()
-            }
-
-            // Duplicate handlers removed
-            Message::View(view::Message::CloseWebview) => {
-                tracing::info!("🌐 [LIANA] Closing webview");
-                #[cfg(feature = "webview")]
-                {
-                    self.webview_mode = false;
-                    self.webview_loading = false;
-                    self.webview_ready = false;
-                    self.show_webview = false;
-                    self.current_webview_url = None;
-                    self.webview_loading_start = None;
-                    self.num_webviews = 0;
-                    self.current_webview_index = None;
-                }
-                Task::none()
-            }
-            #[cfg(feature = "webview")]
-            Message::View(view::Message::ShowWebView) => {
-                tracing::info!("🌐 [LIANA] Showing webview");
-                self.show_webview = true;
-
-                // Create a webview with the current URL if available
-                if let Some(url) = &self.current_webview_url {
-                    tracing::info!("🌐 [LIANA] Creating webview with URL: {}", url);
-                    self.webview_ready = false; // Mark as not ready until view is created
-                    self.webview_loading = true; // Mark as loading
-                    self.webview_loading_start = Some(std::time::Instant::now()); // Track loading start time
-
-                    use iced::Size;
-                    use iced_webview::{Action as WebViewAction, PageType};
-
-                    // Resize webview to match meld container size (800px width, 600px height for better fit)
-                    let webview_size = Size::new(800, 600);
-                    let webview = self.webview.get_or_insert_with(init_webview);
-                    let resize_task = webview
-                        .update(WebViewAction::Resize(webview_size))
-                        .map(Self::map_webview_message_static);
-
-                    // Create the view with the URL - this will trigger the webview to initialize
-                    let create_task = Task::done(Message::View(view::Message::WebviewAction(
-                        WebViewAction::CreateView(PageType::Url(url.clone())),
-                    )));
-
-                    Task::batch([resize_task, create_task])
-                } else {
-                    tracing::warn!("🌐 [LIANA] No URL available for webview");
-                    Task::none()
-                }
-            }
-            #[cfg(feature = "dev-meld")]
-            Message::View(view::Message::MeldBuySell(
-                view::MeldBuySellMessage::SessionCreated(url),
-            )) => {
-                tracing::info!("🌐 [LIANA] Meld session created with URL: {}", url);
-                // Set the URL and show webview immediately - no intermediate button needed
-                self.current_webview_url = Some(url.clone());
-                self.webview_mode = true;
-                self.show_webview = true; // Show webview immediately
-
-                // Initialize loading timer for smart update strategy
-                self.webview_loading_start = Some(std::time::Instant::now());
-                self.webview_loading = true;
-                self.webview_ready = false;
-
-                // Resize webview to match the "Meld Payment Ready" container size
-                // This ensures the webview will be properly sized when shown
-                #[cfg(feature = "webview")]
-                {
-                    use iced::Size;
-                    use iced_webview::Action as WebViewAction;
-
-                    // Set webview size to match the meld container (600px width, 600px height)
-                    let container_size = Size::new(600, 600);
-                    let webview = self.webview.get_or_insert_with(init_webview);
-                    let resize_task = webview
-                        .update(WebViewAction::Resize(container_size))
-                        .map(Self::map_webview_message_static);
-
-                    let panel_update_task = self.panels.current_mut().update(
-                        self.daemon.clone(),
-                        &self.cache,
-                        Message::View(view::Message::MeldBuySell(
-                            view::MeldBuySellMessage::SessionCreated(url),
-                        )),
-                    );
-
-                    Task::batch([resize_task, panel_update_task])
-                }
-                #[cfg(not(feature = "webview"))]
-                {
-                    self.panels.current_mut().update(
-                        self.daemon.clone(),
-                        &self.cache,
-                        Message::View(view::Message::MeldBuySell(
-                            view::MeldBuySellMessage::SessionCreated(url),
-                        )),
-                    )
-                }
-            }
-
+            // TODO: Move to panel.state
             _ => self
                 .panels
                 .current_mut()
@@ -741,101 +458,16 @@ impl App {
     }
 
     pub fn view(&self) -> Element<'_, Message> {
-        // Check if we should show embedded webview within the buy/sell panel
-        #[cfg(feature = "webview")]
-        {
-            if self.webview_mode && matches!(self.panels.current, menu::Menu::BuyAndSell) {
-                // Create webview content that will be embedded below the "Previous" button
-                let webview_widget = {
-                    use crate::app::view::webview::meld_webview_widget_ultralight;
-                    Some(meld_webview_widget_ultralight(
-                        self.webview.as_ref(),
-                        self.current_webview_url.as_deref(),
-                        self.show_webview,
-                        self.webview_ready,
-                        self.webview_loading,
-                        self.current_webview_index,
-                    ))
-                };
+        let view = self.panels.current().view(&self.cache);
 
-                // Use meld buy/sell view with embedded webview for Buy/Sell panel in webview mode
-                let panel_content = {
-                    use crate::app::view::meld_buysell::meld_buysell_view_with_webview;
-                    #[cfg(feature = "dev-meld")]
-                    {
-                        meld_buysell_view_with_webview(
-                            &self.panels.meld_buy_and_sell,
-                            webview_widget,
-                        )
-                    }
-                    #[cfg(not(feature = "dev-coincube"))]
-                    {
-                        // Fallback to normal panel view if meld/onramp features are not enabled
-                        self.panels.current().view(&self.cache)
-                    }
-                };
-
-                // Apply dashboard layout once to maintain left sidebar
-                let dashboard_content =
-                    view::dashboard(&self.panels.current, &self.cache, None, panel_content);
-
-                if self.cache.network != bitcoin::Network::Bitcoin {
-                    Column::with_children(vec![
-                        network_banner(self.cache.network).into(),
-                        dashboard_content.map(Message::View),
-                    ])
-                    .into()
-                } else {
-                    dashboard_content.map(Message::View)
-                }
-            } else {
-                // Normal panel view without webview
-                let panel_content = self.panels.current().view(&self.cache);
-
-                // Buy/Sell panel needs dashboard wrapper, other panels already have it internally
-                let final_view = if matches!(self.panels.current, menu::Menu::BuyAndSell) {
-                    // Apply dashboard wrapper for Buy/Sell panel
-                    view::dashboard(&self.panels.current, &self.cache, None, panel_content)
-                } else {
-                    // Other panels already apply dashboard() internally
-                    panel_content
-                };
-
-                if self.cache.network != bitcoin::Network::Bitcoin {
-                    Column::with_children(vec![
-                        network_banner(self.cache.network).into(),
-                        final_view.map(Message::View),
-                    ])
-                    .into()
-                } else {
-                    final_view.map(Message::View)
-                }
-            }
-        }
-
-        #[cfg(not(feature = "webview"))]
-        {
-            // Normal panel view without webview
-            let panel_content = self.panels.current().view(&self.cache);
-
-            // Buy/Sell panel needs dashboard wrapper, other panels already have it internally
-            let final_view = if matches!(self.panels.current, menu::Menu::BuyAndSell) {
-                // Apply dashboard wrapper for Buy/Sell panel
-                view::dashboard(&self.panels.current, &self.cache, None, panel_content)
-            } else {
-                // Other panels already apply dashboard() internally
-                panel_content
-            };
-
-            if self.cache.network != bitcoin::Network::Bitcoin {
-                Column::with_children(vec![
-                    network_banner(self.cache.network).into(),
-                    final_view.map(Message::View),
-                ])
-                .into()
-            } else {
-                final_view.map(Message::View)
-            }
+        if self.cache.network != bitcoin::Network::Bitcoin {
+            Column::with_children([
+                network_banner(self.cache.network).into(),
+                view.map(Message::View),
+            ])
+            .into()
+        } else {
+            view.map(Message::View)
         }
     }
