+++ liana-gui/src/app/view/mod.rs
@@ -1,14 +1,17 @@
 mod label;
 mod message;
 mod warning;
-pub mod webview;
 
 pub mod coins;
 pub mod export;
 pub mod home;
 pub mod hw;
+
 #[cfg(feature = "dev-meld")]
 pub mod meld_buysell;
+#[cfg(feature = "dev-onramp")]
+pub mod onramper_buysell;
+
 pub mod psbt;
 pub mod psbts;
 pub mod receive;
@@ -138,18 +141,20 @@ pub fn sidebar<'a>(menu: &Menu, cache: &'a Cache) -> Container<'a, Message> {
             .width(iced::Length::Fill))
     };
 
-    let buy_sell_button = if *menu == Menu::BuyAndSell {
-        row!(
-            button::menu_active(Some(bitcoin_icon()), "Buy/Sell")
-                .on_press(Message::Reload)
-                .width(iced::Length::Fill),
-            menu_bar_highlight()
-        )
-    } else {
-        row!(button::menu(Some(bitcoin_icon()), "Buy/Sell")
-            .on_press(Message::Menu(Menu::BuyAndSell))
-            .width(iced::Length::Fill))
-    };
+    let buy_sell_button = cfg!(feature = "dev-coincube").then(|| {
+        if *menu == Menu::BuySell {
+            row!(
+                button::menu_active(Some(bitcoin_icon()), "Buy/Sell")
+                    .on_press(Message::Reload)
+                    .width(iced::Length::Fill),
+                menu_bar_highlight()
+            )
+        } else {
+            row!(button::menu(Some(bitcoin_icon()), "Buy/Sell")
+                .on_press(Message::Menu(Menu::BuySell))
+                .width(iced::Length::Fill))
+        }
+    });
 
     let settings_button = if *menu == Menu::Settings {
         row!(
@@ -180,7 +185,7 @@ pub fn sidebar<'a>(menu: &Menu, cache: &'a Cache) -> Container<'a, Message> {
                     .push(coins_button)
                     .push(transactions_button)
                     .push(psbt_button)
-                    .push(buy_sell_button)
+                    .push_maybe(buy_sell_button)
                     .height(Length::Fill),
             )
             .push(
@@ -295,7 +300,7 @@ pub fn small_sidebar<'a>(menu: &Menu, cache: &'a Cache) -> Container<'a, Message
             .width(iced::Length::Fill))
     };
 
-    let buy_sell_button = if *menu == Menu::BuyAndSell {
+    let buy_sell_button = if *menu == Menu::BuySell {
         row!(
             button::menu_active_small(bitcoin_icon())
                 .on_press(Message::Reload)
@@ -304,7 +309,7 @@ pub fn small_sidebar<'a>(menu: &Menu, cache: &'a Cache) -> Container<'a, Message
         )
     } else {
         row!(button::menu_small(bitcoin_icon())
-            .on_press(Message::Menu(Menu::BuyAndSell))
+            .on_press(Message::Menu(Menu::BuySell))
             .width(iced::Length::Fill))
     };
